requirements.txt
cryptography.fernet
Flask~=3.1.3
Flask-JWT-Extended~=4.7.1
flasgger~=0.9.7.1
psycopg2-binary~=2.9.11
PyYAML~=6.0.3
python-dotenv~=1.2.2
Werkzeug~=3.1.6
Flask-Cors~=5.0.0
cryptography~=46.0.6
firebase-admin